refactor: move `Client` impl blocks out of e2ei module [WPB-15530]
