+++ crypto/src/mls/client/mod.rs
@@ -14,6 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
 
+pub(crate) mod e2e_identity;
 mod epoch_observer;
 mod error;
 pub(crate) mod id;
